+++ NEWS.md
@@ -86,6 +86,13 @@ TODO: sort out what happens to `glance.aov()`
 - Added `tidy.regsubsets()` for best subsets linear regression from the `leaps` package
 
 - Added method `tidy.lm.beta()` to tidy `lm.beta` class models (#545 by @mattle24)
+
+- Add feature for glance.biglm to return df.residual
+
+- Patch bug in glance.lavaan (#577)
+
+- Added tidiers for `drc::drm` models (#574 by @edild)
+
 - `tidy.prcomp()` parameter `matrix` gained new options `"scores"`, `"loadings"`, and `"eigenvalues"` (#557 by @GegznaV)
 
 - `tidy.kmeans()` now uses the names of the input variables in the output by