+++ NEWS.md
@@ -68,6 +68,8 @@ TODO: sort out what happens to `glance.aov()`
 
 ## New tidiers, features and bugfixes
 
+- Added `tidy()` and `glance()` methods for `speedglm` objects from the `speedglm` package
+
 - Return factor rather than numeric class predictions in `.fitted` of `augment.polr()`. (#619)
 
 - `ordinal` tidier rewrite